TASK-86 [Feat]: 대시보드 리스트 호출 사이드바로 이동 & Header 컴포넌트 세분화
