+++ src/components/@shared/Common/AuthHeader.tsx
@@ -1,13 +1,12 @@
 import React, { PropsWithChildren } from 'react';
 
-import { Avatar, Button, Divider, Flex, Group, Text } from '@mantine/core';
+import { Button, Divider, Flex, Group } from '@mantine/core';
 import Image from 'next/image';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
 
-import { useRoot } from '@core/contexts/RootContexts';
-import useDashBoardInfo from '@lib/hooks/useDashBoardInfo';
-import useDevice, { DEVICE } from '@lib/hooks/useDevice';
+import HeaderTitle from './HeaderTitle';
+import MembersProfile from './MembersProfile';
+import UserProfile from './UserProfile';
 
 const HeaderButton = ({ children }: PropsWithChildren) => {
   return (
@@ -17,55 +16,10 @@ const HeaderButton = ({ children }: PropsWithChildren) => {
   );
 };
 
-function getMemberLengthByDevice(device: keyof typeof DEVICE) {
-  if (device === 'desktop') return 4;
-  return 2;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function getMemberMoreLength(arr: any[] | undefined, profileCount: number) {
-  if (!arr) return 0;
-  const moreLength = arr.length - profileCount;
-  if (moreLength < 1) return 0;
-  return moreLength;
-}
-
-function getTitleValue(pathname: string) {
-  if (pathname === '/mydashboard') return '내 대시보드';
-  if (pathname === '/mypage') return '계정 관리';
-  return null;
-}
-
 export default function AuthHeader() {
-  const pathname = usePathname();
-  const device = useDevice();
-  const { user, dashboardid } = useRoot();
-  const { dashBoardMembers, dashBoardDetail } = useDashBoardInfo(dashboardid);
-
-  const profileCount = getMemberLengthByDevice(device);
-  const profileMore = getMemberMoreLength(
-    dashBoardMembers?.members,
-    profileCount
-  );
-  const isManagedPage =
-    pathname.includes('mydashboard') || pathname.includes('mypage');
-  const titleValue = getTitleValue(pathname) || dashBoardDetail?.title;
-
   return (
     <Flex className="fixed left-0 right-0 top-0 z-50 h-[60px] items-center justify-end gap-3 border-b-[1px] border-border-gray bg-white pl-[84px] pr-3 md:h-[70px] md:gap-6 md:px-10 md:pl-[200px] md:pr-10 xl:gap-8 xl:px-[70px] xl:pl-[340px] xl:pr-20">
-      <div
-        className={`grow items-center font-xl-20px-bold xl:flex xl:gap-2 ${isManagedPage ? 'flex' : 'hidden'}`}
-      >
-        <h1>{titleValue}</h1>
-        {!isManagedPage && dashBoardDetail?.createdByMe && (
-          <Image
-            width={20}
-            height={16}
-            src="/icons/crown.png"
-            alt="createByMe"
-          />
-        )}
-      </div>
+      <HeaderTitle />
       <Group className="gap-[6px] font-md-14px-medium">
         <HeaderButton>
           <Image
@@ -88,40 +42,9 @@ export default function AuthHeader() {
           초대하기
         </HeaderButton>
       </Group>
-      {dashBoardMembers?.members && dashBoardMembers?.members.length !== 0 && (
-        <Avatar.Group className="h-[38px]">
-          {dashBoardMembers?.members &&
-            dashBoardMembers?.members.slice(0, profileCount).map(member => (
-              <Avatar key={member.id}>
-                <Image
-                  width={38}
-                  height={38}
-                  src={member.profileImageUrl ?? '/images/small_logo.png'}
-                  alt="member profile"
-                />
-              </Avatar>
-            ))}
-          {profileMore !== 0 && <Avatar>+{profileMore}</Avatar>}
-        </Avatar.Group>
-      )}
+      <MembersProfile />
       <Divider className="bg-border-gray" my="sm" orientation="vertical" />
-      {user && (
-        <Flex className="items-center gap-3">
-          <Avatar>
-            {user.profileImageUrl && (
-              <Image
-                width={38}
-                height={38}
-                src={user.profileImageUrl}
-                alt="my profile"
-              />
-            )}
-          </Avatar>
-          <Text className="hidden font-lg-16px-medium md:block">
-            {user.nickname}
-          </Text>
-        </Flex>
-      )}
+      <UserProfile />
     </Flex>
   );
 }
