+++ src/core/contexts/RootContexts.tsx
@@ -12,11 +12,7 @@ import {
   useState,
 } from 'react';
 
-import { usePathname, useRouter } from 'next/navigation';
-
 import useApi from '@/src/lib/hooks/useApi';
-import { DashboardsResponseDto } from '@core/dtos/DashboardDto';
-import { initialDashboard } from '@lib/constants/initialValues';
 
 import type {
   LoginRequestDto,
@@ -27,28 +23,19 @@ import type {
 interface ContextValue {
   user: UserServiceResponseDto | undefined;
   dashboardid: string | undefined;
-  dashBoardList: Partial<DashboardsResponseDto>;
-
   setDashboardid: Dispatch<SetStateAction<string | undefined>>;
-  redirectDashboard: (id: number) => void;
   login: (body: LoginRequestDto) => Promise<void>;
 }
 
 const RootContext = createContext<ContextValue>({
   user: undefined,
   dashboardid: undefined,
-  dashBoardList: {},
-
   setDashboardid: () => {},
-  redirectDashboard: () => {},
   login: async () => {},
 });
 
 export default function RootProvider({ children }: PropsWithChildren) {
-  const pathname = usePathname();
-  const router = useRouter();
   const [dashboardid, setDashboardid] = useState<string | undefined>(undefined);
-
   const { data: loginData, callApi: postAuthLogin } = useApi<LoginResponseDto>(
     '/auth/login',
     'POST'
@@ -56,16 +43,6 @@ export default function RootProvider({ children }: PropsWithChildren) {
   const { data: user, callApi: getMe } = useApi<UserServiceResponseDto>(
     '/users/me',
     'GET'
-  );
-  const { data: dashBoardList = initialDashboard, callApi: getDashBoardList } =
-    useApi<DashboardsResponseDto>('/dashboards', 'GET');
-
-  const redirectDashboard = useCallback(
-    (id: number) => {
-      setDashboardid(String(id));
-      router.push(`/dashboard/${id}`);
-    },
-    [router]
   );
 
   const login = useCallback(
@@ -81,25 +58,16 @@ export default function RootProvider({ children }: PropsWithChildren) {
       localStorage.setItem('accessToken', loginData?.accessToken);
     }
     getMe(undefined);
-    getDashBoardList(undefined, {
-      params: {
-        navigationMethod: 'infiniteScroll',
-        page: 1,
-        size: 10,
-      },
-    });
-  }, [loginData, getMe, getDashBoardList, router, pathname]);
+  }, [loginData, getMe]);
 
   const value = useMemo(
     () => ({
       user,
       dashboardid,
-      dashBoardList,
       setDashboardid,
       login,
-      redirectDashboard,
     }),
-    [user, dashboardid, login, dashBoardList, redirectDashboard]
+    [user, dashboardid, login]
   );
 
   return <RootContext.Provider value={value}>{children}</RootContext.Provider>;
